pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "file-organizer"
version = "0.1.0"
description = "AI-powered file organizer with semantic classification and undo capabilities"
authors = [{name = "File Organizer Team"}]
requires-python = ">=3.10"
license = {text = "MIT"}
readme = "README.md"
keywords = ["file-organization", "ai", "ml", "clip", "semantic-search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "scikit-learn>=1.3.0",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.3",
    "imagehash>=4.3.1",
    "python-magic-bin>=0.4.14",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tqdm>=4.66.0",
    "structlog>=23.1.0",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "black>=23.7.0",
]

[project.scripts]
file-organizer = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "librosa.*",
    "soundfile.*",
    "imagehash.*",
    "hdbscan.*",
    "umap.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
